refactor(timezone): reorganize timezone imports and utilities
- Removed direct imports of `formatDate`, `formatDateParts`, and `useTimezone` from the timezone module.
- Updated imports to utilize the new formatting utilities from `@/lib/utils/formatting`.
- Cleaned up the timezone settings and related components for better modularity and maintainability.
