+++ src/features/dashboard/members/member-table-row.tsx
@@ -10,12 +10,13 @@ import { PROTECTED_URLS } from '@/configs/urls'
 import type { TeamMember } from '@/core/modules/teams/models'
 import { getTeamDisplayName } from '@/core/modules/teams/utils'
 import { UserAvatar } from '@/features/dashboard/shared'
-import { formatDate, useTimezone } from '@/features/dashboard/timezone'
+import { useTimezone } from '@/features/dashboard/timezone'
 import {
   defaultErrorToast,
   defaultSuccessToast,
   useToast,
 } from '@/lib/hooks/use-toast'
+import { formatDate } from '@/lib/utils/formatting'
 import { useTRPC } from '@/trpc/client'
 import { E2BSquareBadge } from '@/ui/brand'
 import { Avatar, AvatarFallback, AvatarImage } from '@/ui/primitives/avatar'
