+++ src/features/dashboard/sandbox/monitoring/components/chart-overlays.tsx
@@ -1,5 +1,6 @@
-import { formatDate, type Timezone } from '@/features/dashboard/timezone'
+import type { Timezone } from '@/features/dashboard/timezone'
 import { cn } from '@/lib/utils'
+import { formatDate } from '@/lib/utils/formatting'
 import {
   AddIcon,
   BlockIcon,
