+++ src/features/dashboard/settings/keys/api-keys-table-row.tsx
@@ -4,9 +4,10 @@ import { usePostHog } from 'posthog-js/react'
 import { CLI_GENERATED_KEY_NAME } from '@/configs/api'
 import type { TeamAPIKey } from '@/core/modules/keys/models'
 import { IdBadge, UserAvatar } from '@/features/dashboard/shared'
-import { formatDate, useTimezone } from '@/features/dashboard/timezone'
+import { useTimezone } from '@/features/dashboard/timezone'
 import { defaultSuccessToast, useToast } from '@/lib/hooks/use-toast'
 import { cn } from '@/lib/utils'
+import { formatDate } from '@/lib/utils/formatting'
 import { E2BSquareBadge } from '@/ui/brand'
 import { Button } from '@/ui/primitives/button'
 import { KeyIcon, RemoveIcon } from '@/ui/primitives/icons'
