+++ src/features/dashboard/templates/detail/overview/template-section.tsx
@@ -3,8 +3,9 @@
 import type { CSSProperties, ReactNode } from 'react'
 import type { TemplateDetail } from '@/core/modules/templates/models'
 import { getTemplateDisplayName } from '@/features/dashboard/templates/helpers'
-import { formatDateParts, useTimezone } from '@/features/dashboard/timezone'
+import { useTimezone } from '@/features/dashboard/timezone'
 import { useClipboard } from '@/lib/hooks/use-clipboard'
+import { formatDateParts } from '@/lib/utils/formatting'
 import { cn } from '@/lib/utils/ui'
 import { IconButton } from '@/ui/primitives/icon-button'
 import { CheckmarkIcon, CopyIcon } from '@/ui/primitives/icons'
