+++ src/features/dashboard/usage/display-utils.ts
@@ -1,5 +1,9 @@
-import { getDateParts, type Timezone } from '@/features/dashboard/timezone'
-import { formatCurrency, formatNumber } from '@/lib/utils/formatting'
+import type { Timezone } from '@/features/dashboard/timezone'
+import {
+  formatCurrency,
+  formatNumber,
+  getDateParts,
+} from '@/lib/utils/formatting'
 import {
   determineSamplingMode,
   normalizeToEndOfSamplingPeriod,
